sdl2-sys: Skip defining _fltused
Upstream SDL2 will include _fltused in the libary artifact as a
__declspec(selectany) symbol.  This seems to cause problems however with
a definition already coming from the top-most rust link, as it already
gets defined elsewhere when building statically.

Rather than patching the upstream codebase with another patch, just
specify this flag in all cases as we don't require it in any of our rust
builds.